requirements.txt
pydantic~=2.10.0
modbus-tcp-server
jsonpath-ng~=1.7.0
paho-mqtt~=2.1.0
numpy~=2.1.3
click~=8.1.7
psutil~=6.1.0